requirements.txt
# IBM Sterling EDI Application Requirements

# Core dependencies
PyYAML>=6.0.1
requests>=2.31.0

# SQL Server integration
pyodbc>=5.0.0

# API server
flask>=2.3.0
flask-cors>=4.0.0

# Monitoring (optional)
psutil>=5.9.0

# File monitoring (optional, falls back to polling if not available)
watchdog>=3.0.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools (optional)
black>=23.0.0
flake8>=6.1.0

